Add personal databases.

+++ attack/rag_manager.py
@@ -1,20 +1,32 @@
 import os
 
-from langchain.schema import Document
 from langchain_community.vectorstores import FAISS
+from langchain_core.documents import Document
 from langchain_openai import OpenAIEmbeddings
 
-from attack.locations import vector_store_dir
 
 
 class RagManager:
     embeddings: OpenAIEmbeddings
 
-    def __init__(self):
-        self.embeddings = embeddings = OpenAIEmbeddings(api_key=os.environ['OPENAI_API_KEY'])
-        self.db = FAISS.load_local(vector_store_dir, embeddings, allow_dangerous_deserialization=True)
+    vector_store_dir = 'vector_store'
 
-    def retrieve(self, email: str, number_to_retrieve: int = 5):
+    def __init__(self, user: str):
+        self.embeddings = embeddings = OpenAIEmbeddings(api_key=os.environ['OPENAI_API_KEY'])
+        self.user_path = os.path.join(self.vector_store_dir, user)
+        os.makedirs(self.user_path, exist_ok=True)
+
+        file_path = os.path.join(self.user_path, "index.faiss")
+        if not os.path.exists(file_path):
+            self.db = None
+        else:
+           self.db = FAISS.load_local(self.user_path, embeddings, allow_dangerous_deserialization=True)
+
+
+    def retrieve(self, email: str, number_to_retrieve: int = 5) -> list[Document]:
+        if self.db is None:
+            return []
+
         retrieved_rag_docs = self.db.similarity_search(email, k=number_to_retrieve)
         for doc in retrieved_rag_docs:
             print("\nNew Document\n=====================")
@@ -24,7 +36,10 @@ class RagManager:
 
         return retrieved_rag_docs
 
-    def delete_by_id(self, identifiers: list[str]):
+    def delete_by_id(self, identifiers: list[str]) -> list[str]:
+        if self.db is None:
+            return []
+
         success = self.db.delete(identifiers)
         if success is False:
             raise Exception
@@ -32,8 +47,10 @@ class RagManager:
         self.db.save_local(vector_store_dir)
         return identifiers
 
-
     def delete(self, search_phrase: str, deletion_phrase: str, number_to_retrieve: int = 15) -> list[str]:
+        if self.db is None:
+            return []
+
         retrieved = self.retrieve(search_phrase, number_to_retrieve=number_to_retrieve)
 
         document_ids = []
@@ -51,6 +68,10 @@ class RagManager:
 
     def bulk_insert(self, documents: list[Document]) -> None:
         print("inserting documents")
-        self.db.merge_from(FAISS.from_documents(documents, self.embeddings))
-        self.db.save_local(vector_store_dir)
+        if self.db is None:
+            self.db = FAISS.from_documents(documents, self.embeddings)
+        else:
+            self.db.merge_from(FAISS.from_documents(documents, self.embeddings))
+
+        self.db.save_local(self.user_path)
         print("added documents")
